✨:art: made Meldung into a Site form and changed app and router accordingly
